initial commit.
It's been a struggle to implement a custom interator filter that works the way I envision.  I want to be able to pass an argument of what should be matched and use that to build the predicate for filter.
ex.
   something like:
   x.filter_mask(&self,value) -> Filter<Iter<.. {
     self.inner.filter(|f| f == value)
   }

I was not successful.  Instead, the current implementation returns an iterator that has to be assigned to a variable (the call can't be chained - iter won't live long enough), and then filter_mask is called on the var.
ex.
  let mut count = 0;
  let mut z = v.iter_with_mask();
  while let Some(_) = z.filter_mask(&0b00000010) {
      count += 1;
  }
  assert_eq!(count, 3)

I'll keep poking at it.